Make TextSpan u32 to match the compiler one (#1748)
commit-id:107b5e17

---

**Stack**:
- #1749
- #1748 ⬅


⚠️ *Part of a stack created by [spr](https://github.com/ejoffe/spr). Do
not merge manually using the UI - doing so may have unexpected results.*

+++ utils/scarb-test-support/src/proc_macro_server.rs
@@ -41,7 +41,7 @@ pub fn inline_some(token_stream: TokenStream) -> ProcMacroResult {
 #[derive_macro]
 fn some_derive(_token_stream: TokenStream)-> ProcMacroResult {
     let content = "impl SomeImpl of SomeTrait {}".to_string();
-    let span = TextSpan { start: 0, end: content.len() };
+    let span = TextSpan { start: 0, end: content.len() as u32 };
     ProcMacroResult::new(
         TokenStream::new(vec![
             TokenTree::Ident(
